use timer for automatic and manual firing, refactor and comment weapons module
