+++ thetawave_lib/src/player/systems/mod.rs
@@ -1,6 +1,5 @@
 //! Systems for managing players
 
 mod ability;
-mod spaceship;
 
-pub use self::{ability::BarrelRollAbilitySystem, spaceship::SpaceshipSystem};
+pub use self::ability::BarrelRollAbilitySystem;
