--- thetawave_lib/src/weapons/components/mod.rs
@@ -1,5 +1,0 @@
-//! Components used for weapons
-
-mod weapons;
-
-pub use self::weapons::{AutoFireComponent, BlasterComponent, ManualFireComponent};
