--- thetawave_lib/src/weapons/systems/mod.rs
@@ -1,6 +1,0 @@
-//! Systems used for weapons
-
-mod autofire_system;
-mod manualblaster_system;
-
-pub use self::{autofire_system::AutoFireSystem, manualblaster_system::ManualBlasterSystem};
